Changed the name of the function path_frame_init to path_frame_create in class Menu since the other frame creation functions were named in similar way. It is now also in the main_window function like the others.
Created another frame for the initial file selection frames called initial_browse_frame_create because this seems to be the best way to make everything stick to north. I might be wrong about this but will continue experiment with having more frames in frames...
There will probably be at least a 4D frame for when the parallel files selection.
You can also toggle the browse_frame and path_frame now the the toggle_frames function. This function is triggered when the toggle button is pressed(-). This will make it possible to with one line and toggle it and continue with the next one without it taking too much space in the menu.

Added parallel_files_settings_frame_create function in Menu class. This window is for adding and removing parallel lines and the file selection for it. It dosen't have any functionality yet.