Added comment in blf_to_df function in FilesPreperation class to remind me that the two lines below might not be correct.
Changed almost everything with browse in it's name since it wasn't suiting for what it represented. Instead chose line_plot since that's what it really is.

Added a new layer to middle of the window for all the line plots. Needed it to make it stick to the top of the window. In this process I also deleted all the pady logic for file_path_setup and del_path functions in Menu class. This is because of the new approach with layered frames.